requirements.txt
# ============================================================
#  AGI / AI Core Requirements
#  Project: gi (General Intelligence)
# ============================================================

# ── Deep Learning Frameworks ────────────────────────────────
torch>=2.2.0                    # PyTorch — backbone for all neural networks
torchvision>=0.17.0             # Vision transforms & pretrained models
torchaudio>=2.2.0               # Audio processing
tensorflow>=2.16.0              # TensorFlow (optional second backend)
jax[cpu]>=0.4.25                # JAX — functional autodiff, XLA-compiled
flax>=0.8.0                     # Neural networks on top of JAX
optax>=0.2.0                    # Gradient-based optimizers for JAX

# ── Transformers & LLMs ─────────────────────────────────────
transformers>=4.40.0            # HuggingFace — 100k+ pretrained models
tokenizers>=0.19.0              # Fast BPE / WordPiece tokenizers
datasets>=2.19.0                # HuggingFace dataset hub
accelerate>=0.30.0              # Multi-GPU / mixed-precision training
peft>=0.11.0                    # Parameter-Efficient Fine-Tuning (LoRA, etc.)
trl>=0.8.6                      # Transformer Reinforcement Learning (RLHF/PPO)
sentence-transformers>=3.0.0    # Semantic embeddings

# ── Reinforcement Learning ───────────────────────────────────
gymnasium>=1.0.0                # OpenAI Gym successor — RL environments
stable-baselines3>=2.3.0        # PPO, SAC, DQN, A2C implementations
shimmy>=1.3.0                   # Compatibility layer for gym envs
sb3-contrib>=2.3.0              # Extra SB3 algorithms (TQC, TRPO)

# ── Scientific Computing ─────────────────────────────────────
numpy>=1.26.0                   # N-dimensional arrays
scipy>=1.13.0                   # Scientific algorithms
pandas>=2.2.0                   # DataFrames for structured data
sympy>=1.12                     # Symbolic mathematics / reasoning

# ── Memory & Vector Search ───────────────────────────────────
faiss-cpu>=1.8.0                # Facebook AI Similarity Search (vector DB)
chromadb>=0.5.0                 # Persistent vector store
hnswlib>=0.8.0                  # Approximate nearest-neighbor search

# ── Agents & Planning ────────────────────────────────────────
langchain>=0.2.0                # Chain-of-thought agent framework
langchain-community>=0.2.0      # Community integrations for LangChain
langgraph>=0.1.0                # Stateful multi-agent graph execution
openai>=1.30.0                  # OpenAI API (GPT-4o, embeddings)

# ── Neural Architecture Tools ────────────────────────────────
einops>=0.8.0                   # Einstein ops — elegant tensor manipulation
timm>=1.0.0                     # Thousands of pretrained vision models
xformers>=0.0.26                # Memory-efficient attention (from Meta)

# ── Experiment Tracking ──────────────────────────────────────
wandb>=0.17.0                   # Weights & Biases experiment tracker
tensorboard>=2.16.0             # TensorFlow / PyTorch training dashboard
mlflow>=2.13.0                  # ML experiment lifecycle management

# ── Computer Vision ──────────────────────────────────────────
opencv-python>=4.9.0            # Image & video processing
Pillow>=10.3.0                  # Image I/O and transforms
albumentations>=1.4.0           # GPU-accelerated image augmentation

# ── Natural Language Processing ──────────────────────────────
spacy>=3.7.0                    # Industrial NLP pipeline
nltk>=3.8.0                     # Classic NLP toolkit
tiktoken>=0.7.0                 # OpenAI fast BPE tokenizer

# ── Graph Neural Networks ────────────────────────────────────
networkx>=3.3                   # Graph algorithms and data structures
torch-geometric>=2.5.0          # GNNs on PyTorch (relational reasoning)

# ── Visualization ────────────────────────────────────────────
matplotlib>=3.9.0               # Core plotting
seaborn>=0.13.0                 # Statistical visualizations
plotly>=5.22.0                  # Interactive plots

# ── Utilities ────────────────────────────────────────────────
tqdm>=4.66.0                    # Progress bars
pydantic>=2.7.0                 # Data validation & settings
rich>=13.7.0                    # Beautiful terminal output
loguru>=0.7.0                   # Advanced logging
python-dotenv>=1.0.0            # Load .env config files
httpx>=0.27.0                   # Async HTTP client
